Tidied up test
